Take partition size into account in formatting
When formatting the FAT partition on SD cards,
take the partition size into account when choosing
FAT12/FAT16/FAT32 and cluster size. This significantly
reduces memory consumption when fscking bigger cards
prior to mounting.

+++ Fat.h
@@ -26,7 +26,7 @@ public:
                        bool ro, bool remount, bool executable,
                        int ownerUid, int ownerGid, int permMask,
                        bool createLost);
-    static int format(const char *fsPath, unsigned int numSectors);
+    static int format(const char *fsPath, long numSectors);
 };
 
 #endif
